Dedicated test file, invalid follow validation
