+++ solana-programs/anchor/audius-data/lib/lib.ts
@@ -114,8 +114,6 @@ export const initUserSolPubkey = async (args: initUserSolPubkeyArgs) => {
   } = args;
   const signedBytes = signBytes(Buffer.from(message), privateKey);
   const { signature, recoveryId } = signedBytes;
-  console.log(privateKey);
-  console.log(userStgAccount);
   // Get the public key in a compressed format
   const ethPubkey = secp256k1
     .publicKeyCreate(Buffer.from(privateKey, "hex"), false)
@@ -200,22 +198,19 @@ export const createPlaylist = async (args: createPlaylistArgs) => {
     userStgAccountPDA,
     userAuthorityKeypair,
     adminStgPublicKey,
-    metadata
-  } = args;
-  const tx = await program.rpc.createPlaylist(
     metadata,
-    {
-      accounts: {
-        playlist: newPlaylistKeypair.publicKey,
-        user: userStgAccountPDA,
-        authority: userAuthorityKeypair.publicKey,
-        audiusAdmin: adminStgPublicKey,
-        payer: provider.wallet.publicKey,
-        systemProgram: SystemProgram.programId
-      },
-      signers: [newPlaylistKeypair, userAuthorityKeypair]
-    }
-  );
+  } = args;
+  const tx = await program.rpc.createPlaylist(metadata, {
+    accounts: {
+      playlist: newPlaylistKeypair.publicKey,
+      user: userStgAccountPDA,
+      authority: userAuthorityKeypair.publicKey,
+      audiusAdmin: adminStgPublicKey,
+      payer: provider.wallet.publicKey,
+      systemProgram: SystemProgram.programId,
+    },
+    signers: [newPlaylistKeypair, userAuthorityKeypair],
+  });
   return tx;
 };
 
@@ -235,20 +230,17 @@ export const updatePlaylist = async (args: updatePlaylistArgs) => {
     playlistPublicKey,
     userStgAccountPDA,
     userAuthorityKeypair,
-    metadata
-  } = args;
-  const tx = await program.rpc.updatePlaylist(
     metadata,
-    {
-      accounts: {
-        playlist: playlistPublicKey,
-        user: userStgAccountPDA,
-        authority: userAuthorityKeypair.publicKey,
-        payer: provider.wallet.publicKey
-      },
-      signers: [userAuthorityKeypair]
-    }
-  );
+  } = args;
+  const tx = await program.rpc.updatePlaylist(metadata, {
+    accounts: {
+      playlist: playlistPublicKey,
+      user: userStgAccountPDA,
+      authority: userAuthorityKeypair.publicKey,
+      payer: provider.wallet.publicKey,
+    },
+    signers: [userAuthorityKeypair],
+  });
   return tx;
 };
 
@@ -266,18 +258,16 @@ export const deletePlaylist = async (args: deletePlaylistArgs) => {
     program,
     playlistPublicKey,
     userStgAccountPDA,
-    userAuthorityKeypair
+    userAuthorityKeypair,
   } = args;
-  const tx = await program.rpc.deletePlaylist(
-    {
-      accounts: {
-        playlist: playlistPublicKey,
-        user: userStgAccountPDA,
-        authority: userAuthorityKeypair.publicKey,
-        payer: provider.wallet.publicKey
-      },
-      signers: [userAuthorityKeypair]
-    }
-  );
+  const tx = await program.rpc.deletePlaylist({
+    accounts: {
+      playlist: playlistPublicKey,
+      user: userStgAccountPDA,
+      authority: userAuthorityKeypair.publicKey,
+      payer: provider.wallet.publicKey,
+    },
+    signers: [userAuthorityKeypair],
+  });
   return tx;
 };
