+++ solana-programs/anchor/audius-data/tests/audius-data.ts
@@ -334,7 +334,7 @@ describe("audius-data", () => {
         adminStgKeypair,
       });
     } catch (e) {
-      console.log(`ERROR FOUND AS EXPECTED ${e}`);
+      console.log(`Error found as expected ${e}`);
     }
 
     let updatedTrackMetadata = randomCID();
@@ -494,149 +494,5 @@ describe("audius-data", () => {
       }),
     ]);
     console.log(`Created 3 tracks in ${Date.now() - start}ms`);
-  });
-
-  it("Follow/Unfollow user", async () => {
-    // Initialize 2 users
-    let constants1 = initTestConstants();
-    let constants2 = initTestConstants();
-    let handleBytesArray1 = constants1.handleBytesArray;
-    let handleBytesArray2 = constants2.handleBytesArray;
-
-    let { baseAuthorityAccount, bumpSeed, derivedAddress } =
-      await findDerivedPair(
-        program.programId,
-        adminStgKeypair.publicKey,
-        Buffer.from(handleBytesArray1)
-      );
-    let derivedInfo2 = await findDerivedPair(
-      program.programId,
-      adminStgKeypair.publicKey,
-      Buffer.from(handleBytesArray2)
-    );
-
-    let userStorageAccount1 = derivedAddress;
-    let newUserAcct2PDA = derivedInfo2.derivedAddress;
-
-    await testInitUser({
-      provider,
-      program,
-      baseAuthorityAccount,
-      testEthAddr: constants1.testEthAddr,
-      testEthAddrBytes: constants1.testEthAddrBytes,
-      handleBytesArray: handleBytesArray1,
-      bumpSeed,
-      metadata: constants1.metadata,
-      userStgAccount: userStorageAccount1,
-      adminKeypair,
-      adminStgKeypair,
-    });
-
-    await testInitUser({
-      provider,
-      program,
-      baseAuthorityAccount,
-      testEthAddr: constants2.testEthAddr,
-      testEthAddrBytes: constants2.testEthAddrBytes,
-      handleBytesArray: handleBytesArray2,
-      bumpSeed: derivedInfo2.bumpSeed,
-      metadata: constants2.metadata,
-      userStgAccount: newUserAcct2PDA,
-      adminKeypair,
-      adminStgKeypair,
-    });
-
-    // New sol keys that will be used to permission user updates
-    let newUser1Key = anchor.web3.Keypair.generate();
-    let newUser2Key = anchor.web3.Keypair.generate();
-
-    // Generate signed SECP instruction
-    // Message as the incoming public key
-    let message1 = newUser1Key.publicKey.toString();
-    let message2 = newUser2Key.publicKey.toString();
-
-    console.log("initialized");
-
-    await testInitUserSolPubkey({
-      provider,
-      program,
-      message: message1,
-      pkString: constants1.pkString,
-      newUserKeypair: newUser1Key,
-      newUserAcctPDA: userStorageAccount1,
-    });
-    await testInitUserSolPubkey({
-      provider,
-      program,
-      message: message2,
-      pkString: constants2.pkString,
-      newUserKeypair: newUser2Key,
-      newUserAcctPDA: newUserAcct2PDA,
-    });
-
-    // Submit a tx where user 1 follows user 2
-    let followArgs = {
-      accounts: {
-        audiusAdmin: adminStgKeypair.publicKey,
-        payer: provider.wallet.publicKey,
-        authority: newUser1Key.publicKey,
-        followerUserStg: userStorageAccount1,
-        followeeUserStg: newUserAcct2PDA,
-      },
-      signers: [newUser1Key],
-    };
-    let followTx = await program.rpc.followUser(
-      { followUser: {} },
-      handleBytesArray1,
-      handleBytesArray2,
-      followArgs
-    );
-    let txInfo = await confirmLogInTransaction(
-      provider,
-      followTx,
-      "Audius::FollowUser"
-    );
-    const decodedInstruction = program.coder.instruction.decode(
-      txInfo.transaction.message.instructions[0].data,
-      "base58"
-    );
-    // Validate deserialized instructions match input
-    // Confirms user handles passed on chain validation
-    // Can be used during indexing or external tx inspection
-    const instructions = decodedInstruction.data;
-    const user1Handle = String.fromCharCode(...constants1.handleBytesArray);
-    const user2Handle = String.fromCharCode(...constants2.handleBytesArray);
-    const instructionFollowerHandle = String.fromCharCode(
-      ...instructions["followerHandleSeed"]
-    );
-    const instructionFolloweeHandle = String.fromCharCode(
-      ...instructions["followeeHandleSeed"]
-    );
-    assert.equal(user1Handle, instructionFollowerHandle);
-    assert.equal(user2Handle, instructionFolloweeHandle);
-    let unfollowTx = await program.rpc.followUser(
-      { unfollowUser: {} },
-      handleBytesArray1,
-      handleBytesArray2,
-      followArgs
-    );
-    let unFollowtxInfo = await confirmLogInTransaction(
-      provider,
-      unfollowTx,
-      "Audius::UnfollowUser"
-    );
-    const unFollowdecodedInstruction = program.coder.instruction.decode(
-      unFollowtxInfo.transaction.message.instructions[0].data,
-      "base58"
-    );
-    const unfollowInstructions = decodedInstruction.data;
-    const unfInstructionFollowerHandle = String.fromCharCode(
-      ...unfollowInstructions["followerHandleSeed"]
-    );
-    const unfInstructionFolloweeHandle = String.fromCharCode(
-      ...unfollowInstructions["followeeHandleSeed"]
-    );
-    assert.equal(user1Handle, unfInstructionFollowerHandle);
-    assert.equal(user2Handle, unfInstructionFolloweeHandle);
   });
 });
