+++ solana-programs/anchor/audius-data/tests/follows.ts
@@ -1,0 +1,241 @@
+import * as anchor from "@project-serum/anchor";
+import { Program } from "@project-serum/anchor";
+import { assert } from "chai";
+import { initAdmin } from "../lib/lib";
+import { findDerivedPair } from "../lib/utils";
+import { AudiusData } from "../target/types/audius_data";
+import {
+  confirmLogInTransaction,
+  initTestConstants,
+  testInitUser,
+  testInitUserSolPubkey,
+} from "./test-helpers";
+
+const UserActionEnumValues = {
+  unfollowUser: { unfollowUser: {} },
+  followUser: { followUser: {} },
+  invalidEnumValue: { invalidEnum: {} },
+};
+
+describe("follows", () => {
+  const provider = anchor.Provider.local("http://localhost:8899", {
+    preflightCommitment: "confirmed",
+    commitment: "confirmed",
+  });
+
+  // Configure the client to use the local cluster.
+  anchor.setProvider(anchor.Provider.env());
+
+  const program = anchor.workspace.AudiusData as Program<AudiusData>;
+
+  let adminKeypair = anchor.web3.Keypair.generate();
+  let adminStgKeypair = anchor.web3.Keypair.generate();
+
+  it("Initializing admin account!", async () => {
+    await initAdmin({
+      provider: provider,
+      program: program,
+      adminKeypair: adminKeypair,
+      adminStgKeypair: adminStgKeypair,
+      trackIdOffset: new anchor.BN("0"),
+      playlistIdOffset: new anchor.BN("0"),
+    });
+
+    let adminAccount = await program.account.audiusAdmin.fetch(
+      adminStgKeypair.publicKey
+    );
+    if (!adminAccount.authority.equals(adminKeypair.publicKey)) {
+      console.log(
+        "On chain retrieved admin info: ",
+        adminAccount.authority.toString()
+      );
+      console.log("Provided admin info: ", adminKeypair.publicKey.toString());
+      throw new Error("Invalid returned values");
+    }
+  });
+
+  describe("follow / unfollow", () => {
+    let constants1;
+    let constants2;
+    let handleBytesArray1;
+    let handleBytesArray2;
+    // New sol keys that will be used to permission user updates
+    let newUser1Key;
+    let newUser2Key;
+    let userStorageAccount1;
+    let newUserAcct2PDA;
+
+    // Initialize user for each test
+    beforeEach(async () => {
+      // Initialize 2 users
+      constants1 = initTestConstants();
+      constants2 = initTestConstants();
+      handleBytesArray1 = constants1.handleBytesArray;
+      handleBytesArray2 = constants2.handleBytesArray;
+
+      let { baseAuthorityAccount, bumpSeed, derivedAddress } =
+        await findDerivedPair(
+          program.programId,
+          adminStgKeypair.publicKey,
+          Buffer.from(handleBytesArray1)
+        );
+      let derivedInfo2 = await findDerivedPair(
+        program.programId,
+        adminStgKeypair.publicKey,
+        Buffer.from(handleBytesArray2)
+      );
+
+      userStorageAccount1 = derivedAddress;
+      newUserAcct2PDA = derivedInfo2.derivedAddress;
+
+      await testInitUser({
+        provider,
+        program,
+        baseAuthorityAccount,
+        testEthAddr: constants1.testEthAddr,
+        testEthAddrBytes: constants1.testEthAddrBytes,
+        handleBytesArray: handleBytesArray1,
+        bumpSeed,
+        metadata: constants1.metadata,
+        userStgAccount: userStorageAccount1,
+        adminKeypair,
+        adminStgKeypair,
+      });
+
+      await testInitUser({
+        provider,
+        program,
+        baseAuthorityAccount,
+        testEthAddr: constants2.testEthAddr,
+        testEthAddrBytes: constants2.testEthAddrBytes,
+        handleBytesArray: handleBytesArray2,
+        bumpSeed: derivedInfo2.bumpSeed,
+        metadata: constants2.metadata,
+        userStgAccount: newUserAcct2PDA,
+        adminKeypair,
+        adminStgKeypair,
+      });
+
+      // New sol keys that will be used to permission user updates
+      newUser1Key = anchor.web3.Keypair.generate();
+      newUser2Key = anchor.web3.Keypair.generate();
+
+      // Generate signed SECP instruction
+      // Message as the incoming public key
+      let message1 = newUser1Key.publicKey.toString();
+      let message2 = newUser2Key.publicKey.toString();
+
+      await testInitUserSolPubkey({
+        provider,
+        program,
+        message: message1,
+        pkString: constants1.pkString,
+        newUserKeypair: newUser1Key,
+        newUserAcctPDA: userStorageAccount1,
+      });
+      await testInitUserSolPubkey({
+        provider,
+        program,
+        message: message2,
+        pkString: constants2.pkString,
+        newUserKeypair: newUser2Key,
+        newUserAcctPDA: newUserAcct2PDA,
+      });
+    });
+
+    it("follow + unfollow user", async () => {
+      // Submit a tx where user 1 follows user 2
+      let followArgs = {
+        accounts: {
+          audiusAdmin: adminStgKeypair.publicKey,
+          payer: provider.wallet.publicKey,
+          authority: newUser1Key.publicKey,
+          followerUserStg: userStorageAccount1,
+          followeeUserStg: newUserAcct2PDA,
+        },
+        signers: [newUser1Key],
+      };
+      let followTx = await program.rpc.followUser(
+        UserActionEnumValues.followUser,
+        handleBytesArray1,
+        handleBytesArray2,
+        followArgs
+      );
+      let txInfo = await confirmLogInTransaction(
+        provider,
+        followTx,
+        "Audius::FollowUser"
+      );
+      const decodedInstruction = program.coder.instruction.decode(
+        txInfo.transaction.message.instructions[0].data,
+        "base58"
+      );
+      // Validate deserialized instructions match input
+      // Confirms user handles passed on chain validation
+      // Can be used during indexing or external tx inspection
+      const instructions = decodedInstruction.data;
+      const user1Handle = String.fromCharCode(...constants1.handleBytesArray);
+      const user2Handle = String.fromCharCode(...constants2.handleBytesArray);
+      const instructionFollowerHandle = String.fromCharCode(
+        ...instructions["followerHandleSeed"]
+      );
+      const instructionFolloweeHandle = String.fromCharCode(
+        ...instructions["followeeHandleSeed"]
+      );
+      assert.equal(user1Handle, instructionFollowerHandle);
+      assert.equal(user2Handle, instructionFolloweeHandle);
+      let unfollowTx = await program.rpc.followUser(
+        UserActionEnumValues.unfollowUser,
+        handleBytesArray1,
+        handleBytesArray2,
+        followArgs
+      );
+      let unFollowtxInfo = await confirmLogInTransaction(
+        provider,
+        unfollowTx,
+        "Audius::UnfollowUser"
+      );
+      const unFollowdecodedInstruction = program.coder.instruction.decode(
+        unFollowtxInfo.transaction.message.instructions[0].data,
+        "base58"
+      );
+      const unfollowInstructions = decodedInstruction.data;
+      const unfInstructionFollowerHandle = String.fromCharCode(
+        ...unfollowInstructions["followerHandleSeed"]
+      );
+      const unfInstructionFolloweeHandle = String.fromCharCode(
+        ...unfollowInstructions["followeeHandleSeed"]
+      );
+      assert.equal(user1Handle, unfInstructionFollowerHandle);
+      assert.equal(user2Handle, unfInstructionFolloweeHandle);
+    });
+
+    it("submit invalid follow action", async () => {
+      // Submit a tx where user 1 follows user 2
+      let expectedErrorFound = false;
+      let followArgs = {
+        accounts: {
+          audiusAdmin: adminStgKeypair.publicKey,
+          payer: provider.wallet.publicKey,
+          authority: newUser1Key.publicKey,
+          followerUserStg: userStorageAccount1,
+          followeeUserStg: newUserAcct2PDA,
+        },
+        signers: [newUser1Key],
+      };
+      try {
+        // Use invalid enum value and confirm failure
+        await program.rpc.followUser(
+          UserActionEnumValues.invalidEnumValue,
+          handleBytesArray1,
+          handleBytesArray2,
+          followArgs
+        );
+      } catch (e: any) {
+        let index = e.toString().indexOf("unable to infer src variant");
+        if (index > 0) expectedErrorFound = true;
+      }
+      assert.equal(expectedErrorFound, true, "Unable to infer src variant");
+    });
+  });
+});
